Move messages to a different file.
