+++ crates/worker/src/scope.rs
